move download_events_loop to suborder
